pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "compleer"
version = "2025.0.1"
authors = [{ name = "Bertram Aakjær", email = "work.aakjaer@gmail.com" }]
description = "A cross application text completion tool."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "google-genai>=1.56.0",
    "pynput>=1.8.1",
    "python-dotenv>=1.2.1",
    "pywinctl>=0.4.1",
]

[project.scripts]
compleer = "compleer.main:main"

[project.urls]
Homepage = "https://aakjaer.site"


[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q"

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
